Found and fixed hidden ancient bug related to schema reference resolutions.
This bug has been evident when new drawing schemas feature was
implemented, which needs to consume resolved versions of the schemas.

Indeed, this bug has uncovered an obscure bug in jsonschema library,
where it assumes all the "$id" keys are schema or fragment schema
ids, which is not true in the meta-schemas of JSON Schema.

Last, implemented the generation of a graphviz DOT file from a set
of related JSON Schemas.