Major improvements and redesigns in order to lower the memory footprint
The validator memory footprint skyrocketed when lots of validation
errors from several entries were gathered. So, now problematic entries
are yielded as soon as possible, and they are not kept in memory by the
command line program.

Also, the classes and command line program now support iterating over
JSON arrays, so its elements are considered separated entries when it is
indicated. This iteration is done using ijson library, in order to avoid
keeping the whole JSON file in memory.