Version bump to 0.15.0
This version fixes an issue which was uncovered in a very strange corner case.

The issue was related to stale annotations being kept when a JSON schema validation
with guesses is performed. Those stale annotations were being used by the second
pass of the extended validator, and rising "phantom" validation errors.